# Optimize parsing