--- a/src/dynamicprompts/parser/config.py
+++ b/src/dynamicprompts/parser/config.py
@@ -1,6 +1,8 @@
 from dataclasses import dataclass
 
 
+# NB: this class needs to remain `frozen`, because it's used
+#     as a key for `get_cached_parser`'s weak-key dictionary.
 @dataclass(frozen=True)
 class ParserConfig:
     variant_start: str = "{"